test(core): royal flush
remove redundant tests checking the royal cards(10,J,Q,K,A) against every suit to detect a Royal flush. The detection of a flush is tested independetly for all Suits. For this test we want to know that this unique combination (i.e. royal cards + same suit) is detected. It does not matter which suit